requirements.txt
proxmoxer==2.0.*
requests==2.31.*
pythonping==1.1.*
pyyaml==6.0.*
fastapi~=0.111.0
pytest~=8.1.2
pydantic~=2.7.4